feat: Add HyperFocus Zone Master Portal Navigator with mood-based recommendations and portal categories
- Implemented a new HTML page for the HyperFocus Zone Master Portal Navigator.
- Added mood selector buttons to provide tailored portal recommendations based on user feelings.
- Created a responsive design with CSS for an engaging user experience.
- Included JavaScript functionality for mood-based portal suggestions and emergency focus activation.
- Established a comprehensive structure for various portal categories related to focus, community, monetization, support, celebration, and special features.

docs: Create Legendary Immortal Maintenance Activation Complete documentation

- Documented the successful implementation of automated maintenance systems for the empire.
- Outlined the capabilities deployed, achievements unlocked, and a before vs after comparison of operations.
- Highlighted the legendary status achieved with automated processes and monitoring systems.

+++ Python File/ULTRA_THINKING_BOARDROOM_SCANNER.py
@@ -293,8 +293,8 @@ class UltraThinkingBoardroomScanner:
                 error_rate = (total_errors / total_packets) * 100
                 drop_rate = (total_drops / total_packets) * 100
 
-                performance_data["error_rate_percent"] = round(error_rate, 4)
-                performance_data["drop_rate_percent"] = round(drop_rate, 4)
+                performance_data["error_rate_percent"] = float(round(error_rate, 4))
+                performance_data["drop_rate_percent"] = float(round(drop_rate, 4))
 
                 # Health score calculation
                 if error_rate < 0.01 and drop_rate < 0.01:
